+++ nessai/nestedsampler.py
@@ -17,133 +17,21 @@ from tqdm import tqdm
 
 from .livepoint import get_dtype, DEFAULT_FLOAT_DTYPE
 from .plot import plot_indices, plot_trace
-from .posterior import log_integrate_log_trap, logsubexp, LogNegativeError
+from .evidence import (
+    _NSIntegralState,
+    LogNegativeError,
+    logsubexp
+)
 from .proposal import FlowProposal
 from .utils import (
     safe_file_dump,
     compute_indices_ks_test,
-    )
+)
 
 sns.set()
 sns.set_style('ticks')
 
 logger = logging.getLogger(__name__)
-
-
-class _NSIntegralState:
-    """
-    Stores the state of the nested sampling integrator
-
-    Parameters
-    ----------
-    nlive : int
-        Number of live points
-    track_gradients : bool, optional
-        If true the gradient of the change in logL w.r.t logX is saved each
-        time `increment` is called.
-    """
-    def __init__(self, nlive, track_gradients=True):
-        self.nlive = nlive
-        self.reset()
-        self.track_gradients = track_gradients
-
-    @property
-    def log_importance_weights(self):
-        return np.array(self.logLs) + np.array(self.log_vols)
-
-    def reset(self):
-        """
-        Reset the sampler to its initial state at logZ = -infinity
-        """
-        self.logZ = -np.inf
-        self.oldZ = -np.inf
-        self.logX = 0
-        # Information will not be updated on the first iteration, so add
-        # two zeroes so the lists are all consistent
-        self.info = [0., 0.]
-        # Start with a dummy sample enclosing the whole prior
-        self.logLs = [-np.inf]   # Likelihoods sampled
-        self.log_vols = [0.0]    # Volumes enclosed by contours
-        self.gradients = [0]
-
-    def increment(self, x, nlive=None, log_w_norm=None):
-        """
-        Increment the state of the evidence integrator
-        Simply uses rectangle rule for initial estimate
-        """
-        logL = x['logL']
-        logW = x['logW']
-
-        if nlive is None:
-            nlive = self.nlive
-
-        if log_w_norm is None:
-            log_w_norm = np.log(nlive)
-            if not logW == 0.:
-                raise ValueError(
-                    'Weights must be zero when normalisation is None')
-
-        if (logL <= self.logLs[-1]):
-            logger.warning('NS integrator received non-monotonic logL.'
-                           f'{self.logLs[-1]:.5f} -> {logL:.5f}')
-        oldZ = self.logZ
-        logt = -np.exp(logW - log_w_norm)
-        Wt = self.logX + logL + np.log1p(-np.exp(logt))
-        self.logZ = np.logaddexp(self.logZ, Wt)
-        # Update information estimate
-        if np.isfinite(oldZ) and np.isfinite(self.logZ) and np.isfinite(logL):
-            info = np.exp(Wt - self.logZ) * logL \
-                  + np.exp(oldZ - self.logZ) \
-                  * (self.info[-1] + oldZ) \
-                  - self.logZ
-            if np.isnan(info):
-                info = 0
-            self.info.append(info)
-
-        # Update history
-        self.logX += logt
-        self.logLs.append(logL)
-        self.log_vols.append(self.logX)
-        if self.track_gradients:
-            self.gradients.append((self.logLs[-1] - self.logLs[-2])
-                                  / (self.log_vols[-1] - self.log_vols[-2]))
-
-    def finalise(self):
-        """
-        Compute the final evidence with more accurate integrator
-        Call at end of sampling run to refine estimate
-        """
-        # Trapezoidal rule
-        self.logZ = log_integrate_log_trap(np.array(self.logLs),
-                                           np.array(self.log_vols))
-        return self.logZ
-
-    def plot(self, filename=None):
-        """
-        Plot the logX vs logL
-
-        Parameters
-        ----------
-        filename : str, optional
-            Filename name for saving the figure. If not specified the figure
-            is returned.
-        """
-        fig = plt.figure()
-        plt.plot(self.log_vols, self.logLs)
-        plt.title(f'log Z={self.logZ:.2f} '
-                  f'H={self.info[-1] * np.log2(np.e):.2f} bits')
-        plt.grid(which='both')
-        plt.xlabel('log prior-volume')
-        plt.ylabel('log-likelihood')
-        plt.xlim([self.log_vols[-1], self.log_vols[0]])
-        plt.yscale('symlog')
-
-        if filename is not None:
-            fig.savefig(filename, bbox_inches='tight')
-            plt.close()
-            logger.info(f'Saved nested sampling plot as {filename}')
-        else:
-            return fig
 
 
 class NestedSampler: